moved autochooser code into packages
