+++ src/main/java/org/usfirst/frc4048/common/auto/chooser/AutoChooser.java
@@ -1,11 +1,11 @@
-package org.usfirst.frc4048.common.auto;
+package org.usfirst.frc4048.common.auto.chooser;
 
-import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandBase;
+import org.usfirst.frc4048.common.auto.event.AutoEventComparer;
+import org.usfirst.frc4048.common.auto.event.AutoEventProvider;
 
 import java.util.*;
-import java.util.stream.Collectors;
 
 public interface AutoChooser {
      Command getAutoCommand();
