+++ src/main/java/org/usfirst/frc4048/common/auto/chooser/ExampleAutoChooser.java
@@ -1,7 +1,12 @@
-package org.usfirst.frc4048.common.auto;
+package org.usfirst.frc4048.common.auto.chooser;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandBase;
+import org.usfirst.frc4048.common.auto.AutoAction;
+import org.usfirst.frc4048.common.auto.FieldLocation;
+import org.usfirst.frc4048.common.auto.PlaceHolderCommand;
+import org.usfirst.frc4048.common.auto.event.AutoEventComparer;
+import org.usfirst.frc4048.common.auto.event.AutoEventProvider;
 
 import java.util.Map;
