+++ src/main/java/org/usfirst/frc4048/common/auto/event/AutoEvent.java
@@ -1,4 +1,7 @@
-package org.usfirst.frc4048.common.auto;
+package org.usfirst.frc4048.common.auto.event;
+
+import org.usfirst.frc4048.common.auto.AutoAction;
+import org.usfirst.frc4048.common.auto.FieldLocation;
 
 public class AutoEvent {
      private final AutoAction action;
